Refactored sample app GUI for Server
