+++ Linux.Bluetooth.sln
@@ -30,7 +30,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BleTester", "samples\BleTester\BleTester.csproj", "{5794ABD8-6BE8-4878-B4B1-92426E984E49}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleServer", "samples\SampleServer\SampleServer.csproj", "{3F673552-5E1E-4130-852C-036A0E714ED6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleServer", "samples\SampleServer\SampleServer.csproj", "{E2918579-E426-4FF9-87D4-A571C71BA0A5}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "GUI", "GUI", "{1A8CDB9B-703B-48AC-B196-A3A0358E635B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -66,10 +68,10 @@ Global
 		{5794ABD8-6BE8-4878-B4B1-92426E984E49}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5794ABD8-6BE8-4878-B4B1-92426E984E49}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5794ABD8-6BE8-4878-B4B1-92426E984E49}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F673552-5E1E-4130-852C-036A0E714ED6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F673552-5E1E-4130-852C-036A0E714ED6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F673552-5E1E-4130-852C-036A0E714ED6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F673552-5E1E-4130-852C-036A0E714ED6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E2918579-E426-4FF9-87D4-A571C71BA0A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E2918579-E426-4FF9-87D4-A571C71BA0A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E2918579-E426-4FF9-87D4-A571C71BA0A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E2918579-E426-4FF9-87D4-A571C71BA0A5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -80,8 +82,9 @@ Global
 		{52EADF18-B84E-401D-A48C-4397F45CB3A3} = {8DCF476D-3F5A-4228-BE63-B7D6D9E5F629}
 		{DA1BDB4E-CA5D-44D2-B9FD-B77AB9858422} = {5052ABFA-8EB7-4E54-B851-7203081BFB65}
 		{936088A7-5002-4BD1-A1E9-643CB04ACB8D} = {5052ABFA-8EB7-4E54-B851-7203081BFB65}
-		{5794ABD8-6BE8-4878-B4B1-92426E984E49} = {8DCF476D-3F5A-4228-BE63-B7D6D9E5F629}
-		{3F673552-5E1E-4130-852C-036A0E714ED6} = {8DCF476D-3F5A-4228-BE63-B7D6D9E5F629}
+		{5794ABD8-6BE8-4878-B4B1-92426E984E49} = {1A8CDB9B-703B-48AC-B196-A3A0358E635B}
+		{E2918579-E426-4FF9-87D4-A571C71BA0A5} = {1A8CDB9B-703B-48AC-B196-A3A0358E635B}
+		{1A8CDB9B-703B-48AC-B196-A3A0358E635B} = {8DCF476D-3F5A-4228-BE63-B7D6D9E5F629}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {00669A67-D800-45CB-8AF3-2B4EB823A37C}
