--- samples/SampleServer/Directory.Build.props
@@ -1,6 +1,0 @@
-<Project>
-  <PropertyGroup>
-    <Nullable>enable</Nullable>
-    <AvaloniaVersion>11.0.2</AvaloniaVersion>
-  </PropertyGroup>
-</Project>
