+++ samples/SampleServer/Program.cs
@@ -1,24 +1,21 @@
-﻿using System;
-
+using System;
 using Avalonia;
-using Avalonia.ReactiveUI;
 
-namespace SampleServer.Desktop;
+namespace SampleServer;
 
-class Program
+internal sealed class Program
 {
-    // Initialization code. Don't use any Avalonia, third-party APIs or any
-    // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
-    // yet and stuff might break.
-    [STAThread]
-    public static void Main(string[] args) => BuildAvaloniaApp()
-        .StartWithClassicDesktopLifetime(args);
+  // Initialization code. Don't use any Avalonia, third-party APIs or any
+  // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
+  // yet and stuff might break.
+  [STAThread]
+  public static void Main(string[] args) => BuildAvaloniaApp()
+      .StartWithClassicDesktopLifetime(args);
 
-    // Avalonia configuration, don't remove; also used by visual designer.
-    public static AppBuilder BuildAvaloniaApp()
-        => AppBuilder.Configure<App>()
-            .UsePlatformDetect()
-            .WithInterFont()
-            .LogToTrace()
-            .UseReactiveUI();
+  // Avalonia configuration, don't remove; also used by visual designer.
+  public static AppBuilder BuildAvaloniaApp() => AppBuilder
+      .Configure<App>()
+      .UsePlatformDetect()
+      .WithInterFont()
+      .LogToTrace();
 }
