--- samples/SampleServer/ViewModels/MainViewModel.cs
@@ -1,6 +1,0 @@
-﻿namespace SampleServer.ViewModels;
-
-public class MainViewModel : ViewModelBase
-{
-    public string Greeting => "Welcome to Avalonia!";
-}
