--- samples/SampleServer/Views/MainView.axaml.cs
@@ -1,11 +1,0 @@
-﻿using Avalonia.Controls;
-
-namespace SampleServer.Views;
-
-public partial class MainView : UserControl
-{
-    public MainView()
-    {
-        InitializeComponent();
-    }
-}
